Provide a CDI version of MetricsConfiguration
We are trying to make sure this library works for both CDI-only apps
(Quarkus), or for apps that support EJBs also (EAP). To do so, I moved
MetricsConfiguration to MetricsCDIConfiguration, and transformed it from
an EJB managed object to a CDI managed object.

Futhermore, MetricsConfiguration is created for EJB managed object. It
injects MetricsCDIConfiguration. They both expose the same public
methods to make sure the shuffle doesn't break anything on the current
projects using it.
